updated noOfSquares()
added an if statement that will check if "sqr" class present and if
so then will delete. this will help this function as when input
changes it will deleted old grid and output only desired amount of
squares desired by user.
also cleaned up all files out of redundant code and/or adjusted it
to help create the drawing grid.