update noOfSquares(), adding class to cont(s)
I have notice the problem that when slider was adjusted for value
the first part of function would delete divs with id sqrs but not the
div with id cont{i}, which would leave behind unnecessary cluter
and slow down the website.
solved with adding class ="toBeDeleted" and updated arrow function .remove()
to get rid of cluter.