added function to pencil to be able to draw
added onclick function that "assigns" eventListener to each of the
class square with type: 'mouseover' to change background color of
the div into black.
so once button is selected and mouse hovers over the square div it
will change color into black; a "stoke" of mouse will live a line or
nice drawing for the user.